refactor: move participant stats to platform stats
